fix: restructured testing files, moved newtrial to views
