+++ code/app/src/main/AndroidManifest.xml
@@ -27,7 +27,7 @@
         <activity
             android:name=".Views.ViewStatisticActivity"
             android:parentActivityName=".Views.ViewTrialLogActivity" />
-        <activity android:name=".Models.NewTrial" />
+        <activity android:name=".Views.NewTrialActivity" />
         <activity
             android:name=".Views.ViewTrialLogActivity"
             android:label="@string/title_activity_view_trial_log"
