+++ code/app/src/main/res/layout/activity_new_trial.xml
@@ -8,7 +8,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".Models.NewTrial">
+        tools:context=".Views.NewTrialActivity">
 
         <Button
             android:id="@+id/cancelButton"
