+++ code/app/src/test/java/com/cmput301w21t06/crowdfly/Controllers/TrialLogTest.java
@@ -1,4 +1,4 @@
-package com.cmput301w21t06.crowdfly;
+package com.cmput301w21t06.crowdfly.Controllers;
 
 import android.util.Log;
