+++ code/app/src/test/java/com/cmput301w21t06/crowdfly/Models/UserTest.java
@@ -1,4 +1,4 @@
-package com.cmput301w21t06.crowdfly;
+package com.cmput301w21t06.crowdfly.Models;
 
 import com.cmput301w21t06.crowdfly.Models.User;
