+++ code/part3_javadocs/com/cmput301w21t06/crowdfly/BinomialTrialTest.html
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.cmput301w21t06.crowdfly.BinomialTrialTest</li>
+<li>com.cmput301w21t06.crowdfly.Models.BinomialTrialTest</li>
 </ul>
 </li>
 </ul>
