+++ code/part3_javadocs/com/cmput301w21t06/crowdfly/CountTrialTest.html
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.cmput301w21t06.crowdfly.CountTrialTest</li>
+<li>com.cmput301w21t06.crowdfly.Models.CountTrialTest</li>
 </ul>
 </li>
 </ul>
