+++ code/part3_javadocs/com/cmput301w21t06/crowdfly/ExperimentLogTest.html
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.cmput301w21t06.crowdfly.ExperimentLogTest</li>
+<li>com.cmput301w21t06.crowdfly.Controllers.ExperimentLogTest</li>
 </ul>
 </li>
 </ul>
