+++ code/part3_javadocs/com/cmput301w21t06/crowdfly/MeasurementTrialTest.html
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.cmput301w21t06.crowdfly.MeasurementTrialTest</li>
+<li>com.cmput301w21t06.crowdfly.Models.MeasurementTrialTest</li>
 </ul>
 </li>
 </ul>
