+++ code/part3_javadocs/com/cmput301w21t06/crowdfly/Models/NewTrial.html
@@ -123,7 +123,7 @@ var activeTableTab = "activeTableTab";
 <li>androidx.appcompat.app.AppCompatActivity</li>
 <li>
 <ul class="inheritance">
-<li>com.cmput301w21t06.crowdfly.Models.NewTrial</li>
+<li>com.cmput301w21t06.crowdfly.Views.NewTrialActivity</li>
 </ul>
 </li>
 </ul>
